+++ CHANGELOG.md
@@ -74,3 +74,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Integration in Focus and Beam
 - Frontend DTO
 - Email as Mime message
+- EmailSenderIfError annotation